--- operator/pkg/gateway-api/graph/gateway_node.go
@@ -1,92 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright Authors of Cilium
-
-package graph
-
-import (
-	"errors"
-	"sort"
-	"time"
-
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	gatewayv1 "sigs.k8s.io/gateway-api/apis/v1"
-)
-
-func (node *GatewayNode) Validate() error {
-	if ref := node.Gateway.Spec.Infrastructure; ref != nil && ref.ParametersRef != nil {
-		node.SetAccepted(
-			false,
-			"Invalid Gateway parameters: spec.infrastructure.parametersRef is not supported",
-			gatewayv1.GatewayReasonInvalidParameters,
-		)
-		node.SetProgrammed(
-			metav1.ConditionUnknown,
-			"Waiting for Accepted condition to be True",
-			gatewayv1.GatewayReasonPending,
-		)
-		return errors.New("Invalid Gateway")
-	}
-
-	return nil
-}
-
-func (node *GatewayNode) SetAccepted(accepted bool, message string, reason gatewayv1.GatewayConditionReason) {
-	status := metav1.ConditionFalse
-	if accepted {
-		status = metav1.ConditionTrue
-	}
-	node.setCondition(metav1.Condition{
-		Type:               string(gatewayv1.GatewayConditionAccepted),
-		Status:             status,
-		Reason:             string(reason),
-		Message:            message,
-		ObservedGeneration: node.Gateway.GetGeneration(),
-		LastTransitionTime: metav1.NewTime(time.Now()),
-	})
-}
-
-func (node *GatewayNode) SetProgrammed(
-	status metav1.ConditionStatus, message string, reason gatewayv1.GatewayConditionReason,
-) {
-	node.setCondition(metav1.Condition{
-		Type:               string(gatewayv1.GatewayConditionProgrammed),
-		Status:             status,
-		Reason:             string(reason),
-		Message:            message,
-		ObservedGeneration: node.Gateway.GetGeneration(),
-		LastTransitionTime: metav1.NewTime(time.Now()),
-	})
-}
-
-func (node *GatewayNode) setCondition(update metav1.Condition) {
-	for index, condition := range node.Gateway.Status.Conditions {
-		if condition.Type != update.Type {
-			continue
-		}
-		if condition.Status != update.Status ||
-			condition.Reason != update.Reason ||
-			condition.Message != update.Message ||
-			condition.ObservedGeneration != update.ObservedGeneration {
-			node.Gateway.Status.Conditions[index] = update
-		}
-		return
-	}
-
-	node.Gateway.Status.Conditions = append(node.Gateway.Status.Conditions, update)
-}
-
-func (node *GatewayNode) SortListenerSets() {
-	sort.Slice(node.ListenerSets, func(i, j int) bool {
-		left := node.ListenerSets[i].ListenerSet
-		right := node.ListenerSets[j].ListenerSet
-		leftTimestamp := left.CreationTimestamp.Time
-		rightTimestamp := right.CreationTimestamp.Time
-		if !leftTimestamp.Equal(rightTimestamp) {
-			return leftTimestamp.Before(rightTimestamp)
-		}
-
-		leftName := left.GetNamespace() + "/" + left.GetName()
-		rightName := right.GetNamespace() + "/" + right.GetName()
-		return leftName < rightName
-	})
-}
